requirements.txt
asn1crypto==0.24.0
cffi==1.11.4
cryptography==2.1.4
enum34==1.1.6
idna==2.6
ipaddress==1.0.19
pycparser==2.18
pyperclip==1.6.0
PyYAML==3.12
six==1.11.0
